+++ src/components/DatePicker/DatePicker.types.ts
@@ -1,0 +1,4 @@
+export interface Option {
+  value: string
+  label: string
+}
