+++ src/components/DatePicker/components/DayOfWeek.tsx
@@ -1,0 +1,3 @@
+export const DayOfWeek = ({ day }: { day: string }) => (
+  <span className="m-px w-8 block text-center text-gray-500">{day}</span>
+)
