+++ src/components/DatePicker/index.tsx
@@ -2,11 +2,9 @@ import {
   forwardRef,
   ForwardedRef,
   InputHTMLAttributes,
-  CSSProperties,
   useEffect,
   useRef,
   useReducer,
-  useState,
 } from 'react'
 import { ColorShade, ColorPalette } from '../../types'
 import {
@@ -16,17 +14,9 @@ import {
   getLightenColor,
 } from '../../utils'
 import styles from './index.module.css'
-import {
-  getDateFormatByLocale,
-  getDateStringByLocale,
-  getDaysOfWeekByLocale,
-  getMonthsByLocale,
-} from '../../utils/dates'
-import {
-  DatePickerContext,
-  datePickerReducer,
-  useDatePicker,
-} from './use-datepicker'
+import { getDateFormatByLocale } from '../../utils/dates'
+import { DatePickerContext, datePickerReducer } from './use-datepicker'
+import { Calendar, DatePickerInput } from './components'
 
 const createDatePickerStyles = (
   color: string,
@@ -43,453 +33,6 @@ const createDatePickerStyles = (
   '--datepicker-selected-hover-color': selectedHoverColor,
   '--datepicker-error-color': errorColor,
 })
-
-interface Option {
-  value: string
-  label: string
-}
-
-interface DropdownProps {
-  options: Option[]
-  onChange: (selectedValue: string) => void
-  selectedValue: string
-}
-
-function getSelectedIndex(
-  options: Option[],
-  selectedValue: string,
-): number | null {
-  const index = options.findIndex((option) => option.value === selectedValue)
-  return index !== -1 ? index : null
-}
-
-const Dropdown = ({ options, onChange, selectedValue }: DropdownProps) => {
-  const [isOpen, setIsOpen] = useState(false)
-  const [selectedPosition, setSelectedPosition] = useState<number | null>(
-    getSelectedIndex(options, selectedValue),
-  )
-
-  const ref = useRef<HTMLUListElement>(null)
-
-  useEffect(() => {
-    setSelectedPosition(getSelectedIndex(options, selectedValue))
-  }, [options, selectedValue])
-
-  useEffect(() => {
-    const handleOutsideClick = (event: MouseEvent) => {
-      if (ref.current && !ref.current.contains(event.target as Node)) {
-        setIsOpen(false)
-      }
-    }
-
-    if (isOpen) {
-      document.addEventListener('mousedown', handleOutsideClick)
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleOutsideClick)
-    }
-  }, [isOpen])
-
-  useEffect(() => {
-    if (isOpen && selectedPosition !== null && ref.current) {
-      const selectedItem = ref.current.children[
-        selectedPosition
-      ] as HTMLUListElement
-      if (selectedItem) {
-        const scrollOffset = selectedItem.offsetTop - ref.current.offsetTop
-        ref.current.scrollTop = scrollOffset
-      }
-    }
-  }, [isOpen, selectedPosition])
-
-  const handleToggle = () => {
-    setIsOpen(!isOpen)
-    setSelectedPosition(getSelectedIndex(options, selectedValue))
-  }
-
-  const handleSelect = (value: string, index: number) => {
-    onChange(value)
-    setIsOpen(false)
-    setSelectedPosition(index)
-  }
-
-  return (
-    <div className="relative">
-      <button
-        type="button"
-        className="bg-white rounded-md px-2 py-1 text-gray-800 hover:bg-gray-300 focus:outline-none dark:text-gray-200 dark:hover:bg-gray-600 dark:focus:bg-gray-600"
-        onClick={handleToggle}
-      >
-        {options.find((option) => option.value === selectedValue)?.label ||
-          'Select an option'}
-      </button>
-      {isOpen && (
-        <ul
-          ref={ref}
-          className="absolute max-h-40 min-w-fit overflow-y-auto z-10 mt-1 bg-white border border-gray-300 rounded shadow-lg"
-        >
-          {options.map((option, index) => (
-            <li
-              key={option.value}
-              className={`px-4 py-2 cursor-pointer ${
-                selectedValue === option.value ? 'bg-gray-200' : ''
-              }`}
-              onClick={() => handleSelect(option.value, index)}
-            >
-              {option.label}
-            </li>
-          ))}
-        </ul>
-      )}
-    </div>
-  )
-}
-
-const DayOfWeek = ({ day }: { day: string }) => (
-  <span className="m-px w-8 block text-center text-gray-500">{day}</span>
-)
-
-const MonthSelector = ({
-  selectedMonth,
-  months,
-  onChange,
-}: {
-  selectedMonth: number
-  months: string[]
-  onChange: (value: string) => void
-}) => {
-  const monthOptions: Option[] = months.map((month, index) => ({
-    label: month,
-    value: index.toString(),
-  }))
-  return (
-    <Dropdown
-      options={monthOptions}
-      onChange={onChange}
-      selectedValue={selectedMonth.toString()}
-    />
-  )
-}
-
-const YearSelector = ({
-  selectedYear,
-  onChange,
-}: {
-  selectedYear: number
-  onChange: (value: string) => void
-}) => {
-  const currentYear = new Date().getFullYear()
-  const minYear = 1900
-  const maxYear = currentYear + 5
-
-  const yearOptions: Option[] = []
-  for (let year = minYear; year <= maxYear; year++) {
-    yearOptions.push({ label: year?.toString(), value: year?.toString() })
-  }
-
-  return (
-    <Dropdown
-      options={yearOptions}
-      onChange={onChange}
-      selectedValue={selectedYear.toString()}
-    />
-  )
-}
-
-const DatePickerInput = ({
-  className = '',
-  style = {},
-  isRange = false,
-  language = 'en',
-  disabled,
-  onClick,
-  onBlur,
-  ...rest
-}: {
-  onClick: () => void
-  onBlur?: (value?: { startDate?: Date; endDate?: Date }) => void
-  language?: string
-  isRange?: boolean
-  placeholder?: string
-  className?: string
-  style?: CSSProperties
-  disabled?: boolean
-}) => {
-  const {
-    state: { startDate, endDate },
-  } = useDatePicker()
-  const ref = useRef<HTMLInputElement>(null)
-
-  const handleButtonClick = () => {
-    ref.current?.focus()
-    onClick()
-  }
-
-  return (
-    <>
-      <input
-        className={`relative bg-transparent block w-full outline-none text-gray-400 ${className}`}
-        style={style}
-        type="text"
-        onClick={onClick}
-        onBlur={() =>
-          onBlur &&
-          onBlur({
-            ...(startDate && { startDate }),
-            ...(endDate && { endDate }),
-          })
-        }
-        disabled={disabled}
-        defaultValue={getDateStringByLocale({
-          isRange,
-          startDate,
-          endDate,
-          language,
-        })}
-        readOnly
-        ref={ref}
-        {...rest}
-      />
-      <button
-        type="button"
-        className="absolute inset-y-0 right-0 flex items-center px-4 bg-transparent border-transparent"
-        onClick={handleButtonClick}
-        disabled={disabled}
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="24"
-          height="24"
-          viewBox="0 0 24 24"
-          fill="none"
-        >
-          <path
-            d="M8 7V3M16 7V3M7 11H17M5 21H19C20.1046 21 21 20.1046 21 19V7C21 5.89543 20.1046 5 19 5H5C3.89543 5 3 5.89543 3 7V19C3 20.1046 3.89543 21 5 21Z"
-            stroke="#101828"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </svg>
-      </button>
-    </>
-  )
-}
-
-const Calendar = ({
-  language,
-  isSecondCalendar,
-  isRage,
-  minDate,
-  maxDate,
-  onMonthChange,
-  onSelectDate,
-  onYearChange,
-  shouldDisableDate = () => false,
-  alwaysOpen,
-}: {
-  language: string
-  onMonthChange: (value: number) => void
-  onSelectDate: (day: number, isSecondCalendar?: boolean) => void
-  onYearChange: (value: number) => void
-  maxDate?: Date
-  minDate?: Date
-  isSecondCalendar?: boolean
-  isRage?: boolean
-  shouldDisableDate?: (date: Date) => boolean
-  alwaysOpen?: boolean
-}) => {
-  const { state } = useDatePicker()
-
-  const dateButtons = isSecondCalendar
-    ? getCalendarDays(
-        state.secondYear,
-        state.secondMonth,
-        (day) => onSelectDate(day, isSecondCalendar),
-        minDate,
-        maxDate,
-      )
-    : getCalendarDays(state.year, state.month, onSelectDate, minDate, maxDate)
-
-  const handleHoverEffect = (id: string) => {
-    if (!state.modalOpen) return
-
-    const buttons = document.querySelectorAll<HTMLButtonElement>(
-      `.${styles.day}:enabled`,
-    )
-    const selectedButtons = Array.from(buttons).filter((button) =>
-      button.classList.contains(styles.selected),
-    )
-    const hoveredIndex = Array.from(buttons).findIndex(
-      (button) => button.id === id,
-    )
-
-    if (isRage && state.startDate && !state.endDate && selectedButtons[0]) {
-      const selectedIndex = Array.from(buttons).indexOf(selectedButtons[0])
-      buttons.forEach((button, index) => {
-        if (index > selectedIndex && index < hoveredIndex) {
-          button.classList.contains(styles.selected) &&
-            button.classList.remove(styles.selected)
-          button.classList.add(styles.rangeItem)
-        } else if (index === hoveredIndex) {
-          button.classList.add(styles.selected)
-        } else {
-          index !== selectedIndex &&
-            button.classList.remove(styles.rangeItem, styles.selected)
-        }
-      })
-    } else {
-      const selectedIndex = Array.from(buttons).indexOf(selectedButtons[0])
-      const selectedIndexSecond = Array.from(buttons).indexOf(
-        selectedButtons[1],
-      )
-      buttons.forEach((button, index) => {
-        if (
-          selectedButtons.length > 0 &&
-          (selectedIndex === hoveredIndex ||
-            selectedIndexSecond === hoveredIndex) &&
-          index === hoveredIndex
-        ) {
-          button.classList.add(styles.selectedHover)
-        } else if (index === hoveredIndex) {
-          button.classList.add(styles.hovered)
-        } else {
-          button.classList.remove(styles.hovered, styles.selectedHover)
-        }
-      })
-    }
-  }
-
-  function getCalendarDays(
-    year: number,
-    month: number,
-    onSelectDate: (day: number, isSecondCalendar?: boolean) => void,
-    minDate?: Date,
-    maxDate?: Date,
-  ) {
-    const firstDay = new Date(year, month, 1)
-    const lastDay = new Date(year, month + 1, 0)
-    const startDay = firstDay.getDay()
-    const lastDateOfMonth = lastDay.getDate()
-
-    minDate?.setHours(0, 0, 0)
-    maxDate?.setHours(0, 0, 0)
-    const minDateTime = minDate?.getTime() || Number.MIN_VALUE
-    const maxDateTime = maxDate?.getTime() || Number.MAX_VALUE
-
-    const days = Array.from({ length: lastDateOfMonth }, (_, i) => {
-      const day = i + 1
-      const date = new Date(year, month, day)
-      const isCurrentMonth = true
-      const dateTime = date.getTime()
-      const isDisabled =
-        dateTime < minDateTime ||
-        dateTime > maxDateTime ||
-        shouldDisableDate(date)
-      const isToday = day === new Date().getDate() && !isSecondCalendar
-      return { day, date, isCurrentMonth, isDisabled, isToday }
-    })
-
-    const prevMonth = month === 0 ? 11 : month - 1
-    const prevYear = month === 0 ? year - 1 : year
-    const daysInPrevMonth = new Date(prevYear, prevMonth + 1, 0).getDate()
-    const daysFromPrevMonth = startDay === 0 ? 6 : startDay - 1
-
-    const prevMonthDays = Array.from({ length: daysFromPrevMonth }, (_, i) => {
-      const day = daysInPrevMonth - daysFromPrevMonth + i + 1
-      const date = new Date(prevYear, prevMonth, day)
-      const isCurrentMonth = false
-      const isDisabled = true
-      const isToday = false
-      return { day, date, isCurrentMonth, isDisabled, isToday }
-    })
-
-    const nextMonth = month === 11 ? 0 : month + 1
-    const nextYear = month === 11 ? year + 1 : year
-    const daysFromNextMonth = 42 - lastDateOfMonth - daysFromPrevMonth
-
-    const nextMonthDays = Array.from({ length: daysFromNextMonth }, (_, i) => {
-      const day = i + 1
-      const date = new Date(nextYear, nextMonth, day)
-      const isCurrentMonth = false
-      const isDisabled = true
-      const isToday = false
-      return { day, date, isCurrentMonth, isDisabled, isToday }
-    })
-
-    const allDays = [...prevMonthDays, ...days, ...nextMonthDays]
-
-    return allDays.map(({ day, date, isCurrentMonth, isDisabled, isToday }) => {
-      const startDate = state.startDate ? new Date(state.startDate) : null
-      const endDate = state.endDate ? new Date(state.endDate) : null
-
-      const isStartDate =
-        startDate instanceof Date &&
-        isCurrentMonth &&
-        date.toDateString() === startDate.toDateString()
-      const isEndDate =
-        endDate instanceof Date &&
-        isCurrentMonth &&
-        date.toDateString() === endDate.toDateString()
-      const isInRange =
-        state.rangeDays &&
-        state.rangeDays.some(
-          (rangeDate) =>
-            rangeDate instanceof Date &&
-            date.toDateString() === rangeDate.toDateString(),
-        )
-
-      return (
-        <button
-          key={date.toString()}
-          id={date.toString()}
-          type="button"
-          className={`
-            ${styles.day} ${!isCurrentMonth ? 'text-gray-400 cursor-not-allowed' : ''} 
-            ${isStartDate || isEndDate ? `${styles.selected}` : ''} 
-            ${isInRange && !isStartDate && !isEndDate && isCurrentMonth ? `${styles.rangeItem}` : ''}
-            ${isCurrentMonth && isDisabled ? 'line-through' : ''} 
-            ${isToday ? 'text-[--datepicker-scheme]' : ''}
-            ${alwaysOpen || isDisabled ? '' : 'hover:bg-[--datepicker-hover-color]'}
-          `}
-          onClick={() => isCurrentMonth && onSelectDate(day)}
-          onMouseEnter={() =>
-            isCurrentMonth && handleHoverEffect(date.toString())
-          }
-          disabled={isDisabled}
-        >
-          <span>{day}</span>
-        </button>
-      )
-    })
-  }
-
-  return (
-    <div className="p-3">
-      <div className="flex justify-center mb-3">
-        <div className="relative col-span-3 flex justify-center items-center gap-x-1">
-          <MonthSelector
-            selectedMonth={isSecondCalendar ? state.secondMonth : state.month}
-            months={getMonthsByLocale(language)}
-            onChange={(value) => onMonthChange(parseInt(value))}
-          />
-          <span className="text-gray-800 dark:text-gray-200">/</span>
-          <YearSelector
-            selectedYear={isSecondCalendar ? state.secondYear : state.year}
-            onChange={(value) => onYearChange(parseInt(value))}
-          />
-        </div>
-      </div>
-      <div className="flex justify-center gap-x-1 pb-1.5">
-        {getDaysOfWeekByLocale(language).map((day) => (
-          <DayOfWeek key={day} day={day} />
-        ))}
-      </div>
-      <div className="grid grid-cols-7 gap-1">{dateButtons}</div>
-    </div>
-  )
-}
 
 interface DatePickerOwnProps {
   label?: string
